Create repository classes and organized Entity classes
- Create AuthorRepository class.
- Create BookRepository class.
- Create ClientRepository class.
- Create LoanRepository class.
- Organized Repository classes into 'repository' folder
- Organized Entity classes into 'entity' folder
