+++ src/main/java/library/persistence/repository/AuthorRepository.java
@@ -1,0 +1,7 @@
+package library.persistence.repository;
+
+import library.persistence.entity.AuthorEntity;
+import org.springframework.data.repository.CrudRepository;
+
+public interface AuthorRepository extends CrudRepository<AuthorEntity,Long> {
+}
