+++ src/main/java/library/persistence/repository/BookRepository.java
@@ -1,0 +1,7 @@
+package library.persistence.repository;
+
+import library.persistence.entity.BookEntity;
+import org.springframework.data.repository.CrudRepository;
+
+public interface BookRepository extends CrudRepository<BookEntity, Long> {
+}
