+++ src/main/java/library/persistence/repository/ClientRepository.java
@@ -1,0 +1,7 @@
+package library.persistence.repository;
+
+import library.persistence.entity.ClientEntity;
+import org.springframework.data.repository.CrudRepository;
+
+public interface ClientRepository extends CrudRepository<ClientEntity, Long> {
+}
