+++ src/main/java/library/persistence/repository/LoanRepository.java
@@ -1,0 +1,7 @@
+package library.persistence.repository;
+
+import library.persistence.entity.LoanEntity;
+import org.springframework.data.repository.CrudRepository;
+
+public interface LoanRepository extends CrudRepository<LoanEntity,Long> {
+}
